+++ CHANGES.rst
@@ -4,6 +4,11 @@ Changelog
 3.0.5 (unreleased)
 ------------------
 
+- Convert manage-portlets.js into a pattern and make improvements on
+  using the manage portlets infrastructure
+  [vangheem]
+
+
 - Supress a ZopeTestCase warning.
   This fixes https://github.com/plone/Products.CMFPlone/issues/501
   [gforcada]